Add files via upload
This is a project that focuses solely on data cleaning using SQL. This project features a dataset on Housing Data in Tennessee, as someone currently learning the technicalities of data analysis in SQL, I understand the importance of working with data that has been properly cleaned and formatted for analysis. Although this file doesn't contain queries for the main analysis.  I will be sure to update the repository with some actual queries for analysis in SQL.